+++ src/Packages/UnityWebBrowser/Runtime/AssemblyInfo.cs
@@ -15,7 +15,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Voltstro-Studios")]
 [assembly: AssemblyProduct("Unity Web Browser")]
-[assembly: AssemblyCopyright("Copyright (c) 2021-2025 Voltstro-Studios")]
+[assembly: AssemblyCopyright("Copyright (c) 2021-2026 Voltstro-Studios")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -34,8 +34,8 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.2.7")]
-[assembly: AssemblyFileVersion("2.2.7")]
+[assembly: AssemblyVersion("2.2.8")]
+[assembly: AssemblyFileVersion("2.2.8")]
 
 [assembly: InternalsVisibleTo("VoltstroStudios.UnityWebBrowser.Prj")]
 [assembly: InternalsVisibleTo("VoltstroStudios.UnityWebBrowser.Editor")]