+++ src/UnityWebBrowser.Engine.Cef/SubProcess/Program.cs
@@ -16,16 +16,15 @@ public static class Program
     public static int Main(string[] args)
     {
 #if MACOS
-        var sandboxContext = CefSandbox.cef_sandbox_initialize(args.Length, args);
         CefMacOsFrameworkLoader.AddFrameworkLoader();
+        var sandboxContext = CefSandbox.cef_sandbox_initialize(args.Length, args);
 #endif
         
         //Setup CEF
         CefRuntime.Load();
 
         LaunchArgumentsParser argumentsParser = new();
-        int subProcessExitCode = 0;
-        argumentsParser.Run(args, launchArguments =>
+        int processExitCode = argumentsParser.Run(args, launchArguments =>
         {
 // ReSharper disable once RedundantAssignment
             string[] argv = args;
@@ -41,13 +40,13 @@ public static class Program
             UwbCefApp cefApp = new UwbCefApp(launchArguments);
 
             //Run our sub-processes
-            subProcessExitCode = CefRuntime.ExecuteProcess(cefMainArgs, cefApp, IntPtr.Zero);
+            return CefRuntime.ExecuteProcess(cefMainArgs, cefApp, IntPtr.Zero);
         });
 
 #if MACOS
         CefSandbox.cef_sandbox_destroy(sandboxContext);
 #endif
 
-        return subProcessExitCode;
+        return processExitCode;
     }
 }