+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.8] - 2025-01-07
+
+NOTE: Next release is planned to be 3.x breaking change release. The main reason is to drop deprecated code and to up minimal supported Unity version to Unity 6.x.
+
+### Added
+
+- Add configurable WindowlessFrameRate setting (#428)
+
+### Changed
+
+- Update CEF to 143.0.13
+- Bump deps
+- Use custom FixUpdate loop that uses Time.unscaledDeltaTime (#422)
+- Upgrade engine to .NET 10
+
+### Fixed
+
+- Fixed browser not rendering on resize using a workaround for a CEF issue (#425)
+
 ## [2.2.7] - 2025-07-19
 
 ### Changed